removed gevent and grequets as hard dependencies
